Remove isFittable check from ParameterItem
Redmine: #2141
