+++ GUI/coregui/Models/ParameterTreeItems.cpp
@@ -13,19 +13,9 @@
 // ************************************************************************** //
 
 #include "ParameterTreeItems.h"
-#include "BeamItems.h"
-#include "DistributionItems.h"
 #include "FitParameterHelper.h"
-#include "InstrumentItems.h"
-#include "MaterialItem.h"
 #include "ModelPath.h"
-#include "RectangularDetectorItem.h"
 #include "SessionModel.h"
-#include "SphericalDetectorItem.h"
-
-namespace {
-bool containsNames(const QString& text, const QStringList& names);
-}
 
 // ----------------------------------------------------------------------------
 
@@ -82,69 +72,6 @@ void ParameterItem::restoreFromBackup()
     propagateValueToLink(newValue);
 }
 
-//! Returns true if item can be used to drag-and-drop to FitParameterContainer.
-//! In other words, if translation to domain name is implemented and valid.
-
-bool ParameterItem::isFittable() const
-{
-    static const QVector<QPair<QStringList, QStringList>> black_list {
-        {// Global scope
-            {
-                QString()
-            },
-            {
-                Constants::DistributionSigmaFactor,
-                Constants::MaterialRefractiveDataType,
-                Constants::MaterialSLDDataType,
-                MaterialItem::P_MAGNETIZATION
-            }
-        },
-        {// Instrument scope
-            {
-                Constants::GISASInstrumentType,
-                Constants::OffSpecInstrumentType,
-                Constants::SpecularInstrumentType
-            },
-            {// Distribution types
-                Constants::DistributionGateType, Constants::DistributionLorentzType,
-                Constants::DistributionGaussianType, Constants::DistributionLogNormalType,
-                Constants::DistributionCosineType, Constants::DistributionTrapezoidType,
-
-                // Detector axes
-                SphericalDetectorItem::P_PHI_AXIS, SphericalDetectorItem::P_ALPHA_AXIS,
-                RectangularDetectorItem::P_X_AXIS, RectangularDetectorItem::P_Y_AXIS,
-                OffSpecInstrumentItem::P_ALPHA_AXIS,
-
-                // Rectangular detector positioning
-                RectangularDetectorItem::P_ALIGNMENT, RectangularDetectorItem::P_NORMAL,
-                RectangularDetectorItem::P_DIRECTION, RectangularDetectorItem::P_U0,
-                RectangularDetectorItem::P_V0, RectangularDetectorItem::P_DBEAM_U0,
-                RectangularDetectorItem::P_DBEAM_V0, RectangularDetectorItem::P_DISTANCE,
-
-                // Detector resolution
-                Constants::ResolutionFunction2DGaussianType,
-
-                // Beam angle parameters
-                BeamItem::P_INCLINATION_ANGLE, BeamItem::P_AZIMUTHAL_ANGLE
-            }
-        }
-    };
-    Q_ASSERT(ModelPath::ancestor(this, Constants::JobItemType));
-
-    const QString& par_path = FitParameterHelper::getParameterItemPath(this);
-
-    for (const auto& item : black_list) {
-        if (item.first.size() == 1 && item.first[0].isNull()) { // checking global scope
-            if (containsNames(par_path, item.second))
-                return false;
-        } else { // checking everything else
-            if (containsNames(par_path, item.first) && containsNames(par_path, item.second))
-                return false;
-        }
-    }
-    return true;
-}
-
 // ----------------------------------------------------------------------------
 
 ParameterContainerItem::ParameterContainerItem()
@@ -153,14 +80,4 @@ ParameterContainerItem::ParameterContainerItem()
     const QString T_CHILDREN = "children tag";
     registerTag(T_CHILDREN, 0, -1, QStringList() << Constants::ParameterLabelType);
     setDefaultTag(T_CHILDREN);
-}
-
-namespace {
-bool containsNames(const QString& text, const QStringList& names)
-{
-    for (const auto& name : names)
-        if (text.contains(name))
-            return true;
-    return false;
-}
 }
